+++ web/static/react/assets/PracticeProfilePage-BSeKcic2.js
@@ -1,1 +1,0 @@
-import{j as r}from"./vendor-react-DCA3L5Xq.js";import{L as o}from"./LegalSkillsProfilePage-CoEjEXbD.js";import"./button-CEnoOAVa.js";import"./index-Ds8sKtvI.js";import"./vendor-icons-BX2-bB2Y.js";import"./input-Ou5MGn8q.js";import"./IusPageShell-CcWasxlY.js";import"./api-BXSTOx3I.js";function l(){return r.jsx(o,{})}export{l as PracticeProfilePage};